feat: Support all CSS selector styles
Previously, we could only match on prefix strings. With this change,
we allow matching on string suffixes and substring matches too.

This is achieved by introducing a new configuration value "selector",
by default set to Prefix to preserve old behavior. This can be
configured to apply a different CSS selector.

Some work is pending to ensure backwards compatibility, at the moment
this would break existing configurations.